Fix distance/rank type

+++ src/monitor.c
@@ -421,7 +421,7 @@ monitor_t *monitor_from_client(client_t *c)
 
 monitor_t *nearest_monitor(monitor_t *m, direction_t dir, monitor_select_t sel)
 {
-	double dmin = UINT32_MAX;
+	uint32_t dmin = UINT32_MAX;
 	monitor_t *nearest = NULL;
 	xcb_rectangle_t rect = m->rectangle;
 	for (monitor_t *f = mon_head; f != NULL; f = f->next) {
